NIFI-1727: Refactored logic for merging HTTP Requests that are federated across cluster
