+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-cluster/src/main/java/org/apache/nifi/cluster/manager/impl/WebClusterManager.java
@@ -19,26 +19,20 @@ package org.apache.nifi.cluster.manager.impl;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.OutputStream;
 import java.io.Serializable;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
-import java.util.ListIterator;
 import java.util.Map;
-import java.util.NavigableSet;
 import java.util.Queue;
 import java.util.Set;
 import java.util.Timer;
 import java.util.TimerTask;
-import java.util.TreeMap;
-import java.util.TreeSet;
 import java.util.UUID;
 import java.util.concurrent.CompletionService;
 import java.util.concurrent.ConcurrentHashMap;
@@ -54,8 +48,6 @@ import java.util.regex.Pattern;
 
 import javax.net.ssl.SSLContext;
 import javax.ws.rs.HttpMethod;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.StreamingOutput;
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -74,6 +66,8 @@ import org.apache.nifi.annotation.lifecycle.OnRemoved;
 import org.apache.nifi.cluster.HeartbeatPayload;
 import org.apache.nifi.cluster.context.ClusterContext;
 import org.apache.nifi.cluster.context.ClusterContextImpl;
+import org.apache.nifi.cluster.coordination.http.HttpResponseMerger;
+import org.apache.nifi.cluster.coordination.http.StandardHttpResponseMerger;
 import org.apache.nifi.cluster.event.Event;
 import org.apache.nifi.cluster.event.EventManager;
 import org.apache.nifi.cluster.firewall.ClusterNodeFirewall;
@@ -129,8 +123,6 @@ import org.apache.nifi.controller.StandardFlowSerializer;
 import org.apache.nifi.controller.StandardProcessorNode;
 import org.apache.nifi.controller.ValidationContextFactory;
 import org.apache.nifi.controller.exception.ComponentLifeCycleException;
-import org.apache.nifi.controller.queue.DropFlowFileState;
-import org.apache.nifi.controller.queue.ListFlowFileState;
 import org.apache.nifi.controller.reporting.ClusteredReportingTaskNode;
 import org.apache.nifi.controller.reporting.ReportingTaskInstantiationException;
 import org.apache.nifi.controller.reporting.ReportingTaskProvider;
@@ -141,18 +133,8 @@ import org.apache.nifi.controller.scheduling.TimerDrivenSchedulingAgent;
 import org.apache.nifi.controller.service.ControllerServiceLoader;
 import org.apache.nifi.controller.service.ControllerServiceNode;
 import org.apache.nifi.controller.service.ControllerServiceProvider;
-import org.apache.nifi.controller.service.ControllerServiceState;
 import org.apache.nifi.controller.service.StandardControllerServiceProvider;
-import org.apache.nifi.controller.state.SortedStateUtils;
 import org.apache.nifi.controller.state.manager.StandardStateManagerProvider;
-import org.apache.nifi.controller.status.history.ConnectionStatusDescriptor;
-import org.apache.nifi.controller.status.history.MetricDescriptor;
-import org.apache.nifi.controller.status.history.ProcessGroupStatusDescriptor;
-import org.apache.nifi.controller.status.history.ProcessorStatusDescriptor;
-import org.apache.nifi.controller.status.history.RemoteProcessGroupStatusDescriptor;
-import org.apache.nifi.controller.status.history.StandardStatusSnapshot;
-import org.apache.nifi.controller.status.history.StatusHistoryUtil;
-import org.apache.nifi.controller.status.history.StatusSnapshot;
 import org.apache.nifi.encrypt.StringEncryptor;
 import org.apache.nifi.engine.FlowEngine;
 import org.apache.nifi.events.BulletinFactory;
@@ -193,74 +175,9 @@ import org.apache.nifi.util.ReflectionUtils;
 import org.apache.nifi.web.OptimisticLockingManager;
 import org.apache.nifi.web.Revision;
 import org.apache.nifi.web.UpdateRevision;
-import org.apache.nifi.web.api.dto.BulletinBoardDTO;
 import org.apache.nifi.web.api.dto.BulletinDTO;
-import org.apache.nifi.web.api.dto.ComponentStateDTO;
-import org.apache.nifi.web.api.dto.ControllerServiceDTO;
-import org.apache.nifi.web.api.dto.ControllerServiceReferencingComponentDTO;
-import org.apache.nifi.web.api.dto.CountersDTO;
-import org.apache.nifi.web.api.dto.DropRequestDTO;
-import org.apache.nifi.web.api.dto.FlowFileSummaryDTO;
-import org.apache.nifi.web.api.dto.FlowSnippetDTO;
-import org.apache.nifi.web.api.dto.ListingRequestDTO;
-import org.apache.nifi.web.api.dto.NodeCountersSnapshotDTO;
-import org.apache.nifi.web.api.dto.NodeSystemDiagnosticsSnapshotDTO;
-import org.apache.nifi.web.api.dto.ProcessGroupDTO;
-import org.apache.nifi.web.api.dto.ProcessorDTO;
-import org.apache.nifi.web.api.dto.QueueSizeDTO;
-import org.apache.nifi.web.api.dto.RemoteProcessGroupContentsDTO;
-import org.apache.nifi.web.api.dto.RemoteProcessGroupDTO;
-import org.apache.nifi.web.api.dto.RemoteProcessGroupPortDTO;
-import org.apache.nifi.web.api.dto.ReportingTaskDTO;
-import org.apache.nifi.web.api.dto.StateEntryDTO;
-import org.apache.nifi.web.api.dto.StateMapDTO;
-import org.apache.nifi.web.api.dto.SystemDiagnosticsDTO;
-import org.apache.nifi.web.api.dto.provenance.ProvenanceDTO;
-import org.apache.nifi.web.api.dto.provenance.ProvenanceEventDTO;
-import org.apache.nifi.web.api.dto.provenance.ProvenanceRequestDTO;
-import org.apache.nifi.web.api.dto.provenance.ProvenanceResultsDTO;
-import org.apache.nifi.web.api.dto.status.ConnectionStatusDTO;
 import org.apache.nifi.web.api.dto.status.ControllerStatusDTO;
-import org.apache.nifi.web.api.dto.status.NodeConnectionStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.NodePortStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.NodeProcessGroupStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.NodeProcessorStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.NodeRemoteProcessGroupStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.NodeStatusSnapshotsDTO;
-import org.apache.nifi.web.api.dto.status.PortStatusDTO;
-import org.apache.nifi.web.api.dto.status.ProcessGroupStatusDTO;
-import org.apache.nifi.web.api.dto.status.ProcessGroupStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.ProcessorStatusDTO;
-import org.apache.nifi.web.api.dto.status.RemoteProcessGroupStatusDTO;
-import org.apache.nifi.web.api.dto.status.RemoteProcessGroupStatusSnapshotDTO;
-import org.apache.nifi.web.api.dto.status.StatusHistoryDTO;
-import org.apache.nifi.web.api.dto.status.StatusSnapshotDTO;
-import org.apache.nifi.web.api.entity.BulletinBoardEntity;
-import org.apache.nifi.web.api.entity.ComponentStateEntity;
-import org.apache.nifi.web.api.entity.ConnectionStatusEntity;
-import org.apache.nifi.web.api.entity.ControllerServiceEntity;
-import org.apache.nifi.web.api.entity.ControllerServiceReferencingComponentsEntity;
-import org.apache.nifi.web.api.entity.ControllerServicesEntity;
 import org.apache.nifi.web.api.entity.ControllerStatusEntity;
-import org.apache.nifi.web.api.entity.CountersEntity;
-import org.apache.nifi.web.api.entity.DropRequestEntity;
-import org.apache.nifi.web.api.entity.FlowSnippetEntity;
-import org.apache.nifi.web.api.entity.ListingRequestEntity;
-import org.apache.nifi.web.api.entity.PortStatusEntity;
-import org.apache.nifi.web.api.entity.ProcessGroupEntity;
-import org.apache.nifi.web.api.entity.ProcessGroupStatusEntity;
-import org.apache.nifi.web.api.entity.ProcessorEntity;
-import org.apache.nifi.web.api.entity.ProcessorStatusEntity;
-import org.apache.nifi.web.api.entity.ProcessorsEntity;
-import org.apache.nifi.web.api.entity.ProvenanceEntity;
-import org.apache.nifi.web.api.entity.ProvenanceEventEntity;
-import org.apache.nifi.web.api.entity.RemoteProcessGroupEntity;
-import org.apache.nifi.web.api.entity.RemoteProcessGroupStatusEntity;
-import org.apache.nifi.web.api.entity.RemoteProcessGroupsEntity;
-import org.apache.nifi.web.api.entity.ReportingTaskEntity;
-import org.apache.nifi.web.api.entity.ReportingTasksEntity;
-import org.apache.nifi.web.api.entity.StatusHistoryEntity;
-import org.apache.nifi.web.api.entity.SystemDiagnosticsEntity;
 import org.apache.nifi.web.util.WebUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -329,63 +246,16 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
     private static final int DEFAULT_CONNECTION_REQUEST_TRY_AGAIN_SECONDS = 5;
 
 
-    public static final Pattern PROCESSORS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/processors");
-    public static final Pattern PROCESSOR_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/processors/[a-f0-9\\-]{36}");
-    public static final Pattern PROCESSOR_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/processors/[a-f0-9\\-]{36}/status");
-    public static final Pattern PROCESSOR_STATE_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/processors/[a-f0-9\\-]{36}/state");
     public static final Pattern CLUSTER_PROCESSOR_URI_PATTERN = Pattern.compile("/nifi-api/cluster/processors/[a-f0-9\\-]{36}");
 
-    public static final Pattern REMOTE_PROCESS_GROUPS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/remote-process-groups");
-    public static final Pattern REMOTE_PROCESS_GROUP_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/remote-process-groups/[a-f0-9\\-]{36}");
 
-    public static final Pattern PROCESS_GROUP_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))");
-    public static final Pattern GROUP_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/status");
-    public static final Pattern CONTROLLER_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/status");
-    public static final Pattern TEMPLATE_INSTANCE_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/template-instance");
-    public static final Pattern FLOW_SNIPPET_INSTANCE_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/snippet-instance");
-
-    public static final String PROVENANCE_URI = "/nifi-api/controller/provenance";
-    public static final Pattern PROVENANCE_QUERY_URI = Pattern.compile("/nifi-api/controller/provenance/[a-f0-9\\-]{36}");
-    public static final Pattern PROVENANCE_EVENT_URI = Pattern.compile("/nifi-api/controller/provenance/events/[0-9]+");
-
-    public static final String CONTROLLER_SERVICES_URI = "/nifi-api/controller/controller-services/node";
-    public static final Pattern CONTROLLER_SERVICE_URI_PATTERN = Pattern.compile("/nifi-api/controller/controller-services/node/[a-f0-9\\-]{36}");
-    public static final Pattern CONTROLLER_SERVICE_STATE_URI_PATTERN = Pattern.compile("/nifi-api/controller/controller-services/node/[a-f0-9\\-]{36}/state");
-    public static final Pattern CONTROLLER_SERVICE_REFERENCES_URI_PATTERN = Pattern.compile("/nifi-api/controller/controller-services/node/[a-f0-9\\-]{36}/references");
-    public static final String REPORTING_TASKS_URI = "/nifi-api/controller/reporting-tasks/node";
     public static final Pattern REPORTING_TASK_URI_PATTERN = Pattern.compile("/nifi-api/controller/reporting-tasks/node/[a-f0-9\\-]{36}");
-    public static final Pattern REPORTING_TASK_STATE_URI_PATTERN = Pattern.compile("/nifi-api/controller/reporting-tasks/node/[a-f0-9\\-]{36}/state");
-    public static final Pattern BULLETIN_BOARD_URI_PATTERN = Pattern.compile("/nifi-api/controller/bulletin-board");
-    public static final Pattern SYSTEM_DIAGNOSTICS_URI_PATTERN = Pattern.compile("/nifi-api/system-diagnostics");
-    public static final Pattern COUNTERS_URI_PATTERN = Pattern.compile("/nifi-api/controller/counters");
     public static final Pattern COUNTER_URI_PATTERN = Pattern.compile("/nifi-api/controller/counters/[a-f0-9\\-]{36}");
 
-    public static final Pattern PROCESSOR_STATUS_HISTORY_URI_PATTERN =
-        Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/processors/[a-f0-9\\-]{36}/status/history");
-    public static final Pattern PROCESS_GROUP_STATUS_HISTORY_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/status/history");
-    public static final Pattern REMOTE_PROCESS_GROUP_STATUS_HISTORY_URI_PATTERN = Pattern
-        .compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/remote-process-groups/[a-f0-9\\-]{36}/status/history");
-    public static final Pattern CONNECTION_STATUS_HISTORY_URI_PATTERN = Pattern
-        .compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/status/history");
 
-    public static final Pattern CONNECTION_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/status");
-    public static final Pattern INPUT_PORT_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/input-ports/[a-f0-9\\-]{36}/status");
-    public static final Pattern OUTPUT_PORT_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/output-ports/[a-f0-9\\-]{36}/status");
-    public static final Pattern REMOTE_PROCESS_GROUP_STATUS_URI_PATTERN =
-        Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/remote-process-groups/[a-f0-9\\-]{36}/status");
-
-    @Deprecated
-    public static final Pattern QUEUE_CONTENTS_URI = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/contents");
-    public static final Pattern DROP_REQUESTS_URI = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/drop-requests");
-    public static final Pattern DROP_REQUEST_URI = Pattern.compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/drop-requests/[a-f0-9\\-]{36}");
-    public static final Pattern LISTING_REQUESTS_URI = Pattern
-        .compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/listing-requests");
-    public static final Pattern LISTING_REQUEST_URI = Pattern
-        .compile("/nifi-api/controller/process-groups/(?:(?:root)|(?:[a-f0-9\\-]{36}))/connections/[a-f0-9\\-]{36}/listing-requests/[a-f0-9\\-]{36}");
 
     private final NiFiProperties properties;
     private final HttpRequestReplicator httpRequestReplicator;
-    private final HttpResponseMapper httpResponseMapper;
     private final DataFlowManagementService dataFlowManagementService;
     private final ClusterManagerProtocolSenderListener senderListener;
     private final OptimisticLockingManager optimisticLockingManager;
@@ -417,8 +287,8 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
     private final FlowEngine reportingTaskEngine;
     private final StandardProcessScheduler processScheduler;
     private final StateManagerProvider stateManagerProvider;
-    private final long componentStatusSnapshotMillis;
 
+    private final HttpResponseMerger responseMerger = new StandardHttpResponseMerger();
 
     public WebClusterManager(final HttpRequestReplicator httpRequestReplicator, final HttpResponseMapper httpResponseMapper,
             final DataFlowManagementService dataFlowManagementService, final ClusterManagerProtocolSenderListener senderListener,
@@ -438,7 +308,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
 
         // Ensure that our encryptor/decryptor is properly initialized
         this.httpRequestReplicator = httpRequestReplicator;
-        this.httpResponseMapper = httpResponseMapper;
         this.dataFlowManagementService = dataFlowManagementService;
         this.properties = properties;
         this.bulletinRepository = new VolatileBulletinRepository();
@@ -448,15 +317,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         this.optimisticLockingManager = optimisticLockingManager;
         senderListener.addHandler(this);
         senderListener.setBulletinRepository(bulletinRepository);
-
-        final String snapshotFrequency = properties.getProperty(NiFiProperties.COMPONENT_STATUS_SNAPSHOT_FREQUENCY, NiFiProperties.DEFAULT_COMPONENT_STATUS_SNAPSHOT_FREQUENCY);
-        long snapshotMillis;
-        try {
-            snapshotMillis = FormatUtils.getTimeDuration(snapshotFrequency, TimeUnit.MILLISECONDS);
-        } catch (final Exception e) {
-            snapshotMillis = FormatUtils.getTimeDuration(NiFiProperties.DEFAULT_COMPONENT_STATUS_SNAPSHOT_FREQUENCY, TimeUnit.MILLISECONDS);
-        }
-        componentStatusSnapshotMillis = snapshotMillis;
 
         remoteInputPort = properties.getRemoteInputPort();
         if (remoteInputPort == null) {
@@ -2287,482 +2147,16 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         return holder.get();
     }
 
-    private static boolean isProcessorsEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROCESSORS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isProcessorEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method))
-                && (PROCESSOR_URI_PATTERN.matcher(uri.getPath()).matches() || CLUSTER_PROCESSOR_URI_PATTERN.matcher(uri.getPath()).matches())) {
-            return true;
-        } else if ("POST".equalsIgnoreCase(method) && PROCESSORS_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static boolean isProcessorStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROCESSOR_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isProcessorStateEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROCESSOR_STATE_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isProcessGroupEndpoint(final URI uri, final String method) {
-        return ("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method)) && PROCESS_GROUP_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isConnectionStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && CONNECTION_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isInputPortStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && INPUT_PORT_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isOutputPortStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && OUTPUT_PORT_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isRemoteProcessGroupStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && REMOTE_PROCESS_GROUP_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isGroupStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && GROUP_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isControllerStatusEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && CONTROLLER_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isTemplateEndpoint(final URI uri, final String method) {
-        return "POST".equalsIgnoreCase(method) && TEMPLATE_INSTANCE_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isFlowSnippetEndpoint(final URI uri, final String method) {
-        return "POST".equalsIgnoreCase(method) && FLOW_SNIPPET_INSTANCE_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isRemoteProcessGroupsEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && REMOTE_PROCESS_GROUPS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isProcessorStatusHistoryEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROCESSOR_STATUS_HISTORY_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isProcessGroupStatusHistoryEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROCESS_GROUP_STATUS_HISTORY_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isRemoteProcessGroupStatusHistoryEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && REMOTE_PROCESS_GROUP_STATUS_HISTORY_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isConnectionStatusHistoryEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && CONNECTION_STATUS_HISTORY_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isBulletinBoardEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && BULLETIN_BOARD_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isSystemDiagnosticsEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && SYSTEM_DIAGNOSTICS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isCountersEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && COUNTERS_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-
-    private static boolean isRemoteProcessGroupEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method)) && REMOTE_PROCESS_GROUP_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if ("POST".equalsIgnoreCase(method) && REMOTE_PROCESS_GROUPS_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static boolean isProvenanceQueryEndpoint(final URI uri, final String method) {
-        if ("POST".equalsIgnoreCase(method) && PROVENANCE_URI.equals(uri.getPath())) {
-            return true;
-        } else if ("GET".equalsIgnoreCase(method) && PROVENANCE_QUERY_URI.matcher(uri.getPath()).matches()) {
-            return true;
-        }
-        return false;
-    }
-
-    private static boolean isProvenanceEventEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && PROVENANCE_EVENT_URI.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isListFlowFilesEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "DELETE".equalsIgnoreCase(method)) && LISTING_REQUEST_URI.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if ("POST".equalsIgnoreCase(method) && LISTING_REQUESTS_URI.matcher(uri.getPath()).matches()) {
-            return true;
-        }
-
-        return false;
-    }
 
     private static boolean isCounterEndpoint(final URI uri) {
         return COUNTER_URI_PATTERN.matcher(uri.getPath()).matches();
     }
 
-    private static boolean isControllerServicesEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && CONTROLLER_SERVICES_URI.equals(uri.getPath());
-    }
-
-    private static boolean isControllerServiceEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method)) && CONTROLLER_SERVICE_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if ("POST".equalsIgnoreCase(method) && CONTROLLER_SERVICES_URI.equals(uri.getPath())) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static boolean isControllerServiceStateEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && CONTROLLER_SERVICE_STATE_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isControllerServiceReferenceEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method)) && CONTROLLER_SERVICE_REFERENCES_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static boolean isReportingTasksEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && REPORTING_TASKS_URI.equals(uri.getPath());
-    }
-
-    private static boolean isReportingTaskEndpoint(final URI uri, final String method) {
-        if (("GET".equalsIgnoreCase(method) || "PUT".equalsIgnoreCase(method)) && REPORTING_TASK_URI_PATTERN.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if ("POST".equalsIgnoreCase(method) && REPORTING_TASKS_URI.equals(uri.getPath())) {
-            return true;
-        }
-
-        return false;
-    }
-
-    private static boolean isReportingTaskStateEndpoint(final URI uri, final String method) {
-        return "GET".equalsIgnoreCase(method) && REPORTING_TASK_STATE_URI_PATTERN.matcher(uri.getPath()).matches();
-    }
-
-    private static boolean isDropRequestEndpoint(final URI uri, final String method) {
-        if ("DELETE".equalsIgnoreCase(method) && QUEUE_CONTENTS_URI.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if (("GET".equalsIgnoreCase(method) || "DELETE".equalsIgnoreCase(method)) && DROP_REQUEST_URI.matcher(uri.getPath()).matches()) {
-            return true;
-        } else if (("POST".equalsIgnoreCase(method) && DROP_REQUESTS_URI.matcher(uri.getPath()).matches())) {
-            return true;
-        }
-
-        return false;
-    }
 
     static boolean isResponseInterpreted(final URI uri, final String method) {
-        return isProcessorsEndpoint(uri, method) || isProcessorEndpoint(uri, method) || isProcessorStateEndpoint(uri, method)
-                || isRemoteProcessGroupsEndpoint(uri, method) || isRemoteProcessGroupEndpoint(uri, method)
-                || isProcessGroupEndpoint(uri, method)
-                || isTemplateEndpoint(uri, method) || isFlowSnippetEndpoint(uri, method)
-                || isProvenanceQueryEndpoint(uri, method) || isProvenanceEventEndpoint(uri, method)
-                || isControllerServicesEndpoint(uri, method) || isControllerServiceEndpoint(uri, method)
-                || isControllerServiceReferenceEndpoint(uri, method) || isControllerServiceStateEndpoint(uri, method)
-                || isReportingTasksEndpoint(uri, method) || isReportingTaskEndpoint(uri, method) || isReportingTaskStateEndpoint(uri, method)
-                || isDropRequestEndpoint(uri, method) || isListFlowFilesEndpoint(uri, method)
-                || isGroupStatusEndpoint(uri, method) || isProcessorStatusEndpoint(uri, method) || isControllerStatusEndpoint(uri, method)
-                || isConnectionStatusEndpoint(uri, method) || isRemoteProcessGroupStatusEndpoint(uri, method)
-                || isInputPortStatusEndpoint(uri, method) || isOutputPortStatusEndpoint(uri, method)
-                || isProcessorStatusHistoryEndpoint(uri, method) || isProcessGroupStatusHistoryEndpoint(uri, method)
-                || isRemoteProcessGroupStatusHistoryEndpoint(uri, method) || isConnectionStatusHistoryEndpoint(uri, method)
-                || isBulletinBoardEndpoint(uri, method) || isSystemDiagnosticsEndpoint(uri, method)
-                || isCountersEndpoint(uri, method);
+        return StandardHttpResponseMerger.isResponseInterpreted(uri, method);
     }
 
-    private void mergeProcessorValidationErrors(final ProcessorDTO processor, Map<NodeIdentifier, ProcessorDTO> processorMap) {
-        final Map<String, Set<NodeIdentifier>> validationErrorMap = new HashMap<>();
-
-        for (final Map.Entry<NodeIdentifier, ProcessorDTO> nodeEntry : processorMap.entrySet()) {
-            final NodeIdentifier nodeId = nodeEntry.getKey();
-            final ProcessorDTO nodeProcessor = nodeEntry.getValue();
-
-            // merge the validation errors
-            mergeValidationErrors(validationErrorMap, nodeId, nodeProcessor.getValidationErrors());
-        }
-
-        // set the merged the validation errors
-        processor.setValidationErrors(normalizedMergedValidationErrors(validationErrorMap, processorMap.size()));
-    }
-
-    private void mergeComponentState(final ComponentStateDTO componentState, Map<NodeIdentifier, ComponentStateDTO> componentStateMap) {
-        List<StateEntryDTO> localStateEntries = new ArrayList<>();
-
-        int totalStateEntries = 0;
-        for (final Map.Entry<NodeIdentifier, ComponentStateDTO> nodeEntry : componentStateMap.entrySet()) {
-            final ComponentStateDTO nodeComponentState = nodeEntry.getValue();
-            final NodeIdentifier nodeId = nodeEntry.getKey();
-            final String nodeAddress = nodeId.getApiAddress() + ":" + nodeId.getApiPort();
-
-            final StateMapDTO nodeLocalStateMap = nodeComponentState.getLocalState();
-            if (nodeLocalStateMap.getState() != null) {
-                totalStateEntries += nodeLocalStateMap.getTotalEntryCount();
-
-                for (final StateEntryDTO nodeStateEntry : nodeLocalStateMap.getState()) {
-                    nodeStateEntry.setClusterNodeId(nodeId.getId());
-                    nodeStateEntry.setClusterNodeAddress(nodeAddress);
-                    localStateEntries.add(nodeStateEntry);
-                }
-            }
-        }
-
-        // ensure appropriate sort
-        Collections.sort(localStateEntries, SortedStateUtils.getEntryDtoComparator());
-
-        // sublist if necessary
-        if (localStateEntries.size() > SortedStateUtils.MAX_COMPONENT_STATE_ENTRIES) {
-            localStateEntries = localStateEntries.subList(0, SortedStateUtils.MAX_COMPONENT_STATE_ENTRIES);
-        }
-
-        // add all the local state entries
-        componentState.getLocalState().setTotalEntryCount(totalStateEntries);
-        componentState.getLocalState().setState(localStateEntries);
-    }
-
-
-    private void mergeSystemDiagnostics(final SystemDiagnosticsDTO target, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, SystemDiagnosticsDTO> resultMap) {
-        final SystemDiagnosticsDTO mergedSystemDiagnostics = target;
-        mergedSystemDiagnostics.setNodeSnapshots(new ArrayList<NodeSystemDiagnosticsSnapshotDTO>());
-
-        final NodeSystemDiagnosticsSnapshotDTO selectedNodeSnapshot = new NodeSystemDiagnosticsSnapshotDTO();
-        selectedNodeSnapshot.setSnapshot(target.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedSystemDiagnostics.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        for (final Map.Entry<NodeIdentifier, SystemDiagnosticsDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final SystemDiagnosticsDTO toMerge = entry.getValue();
-            if (toMerge == target) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedSystemDiagnostics, toMerge, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergeCounters(final CountersDTO target, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, CountersDTO> resultMap) {
-        final CountersDTO mergedCounters = target;
-        mergedCounters.setNodeSnapshots(new ArrayList<NodeCountersSnapshotDTO>());
-
-        final NodeCountersSnapshotDTO selectedNodeSnapshot = new NodeCountersSnapshotDTO();
-        selectedNodeSnapshot.setSnapshot(target.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedCounters.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        for (final Map.Entry<NodeIdentifier, CountersDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final CountersDTO toMerge = entry.getValue();
-            if (toMerge == target) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedCounters, toMerge, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergeGroupStatus(final ProcessGroupStatusDTO statusDto, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, ProcessGroupStatusDTO> resultMap) {
-        final ProcessGroupStatusDTO mergedProcessGroupStatus = statusDto;
-        mergedProcessGroupStatus.setNodeSnapshots(new ArrayList<NodeProcessGroupStatusSnapshotDTO>());
-
-        final NodeProcessGroupStatusSnapshotDTO selectedNodeSnapshot = new NodeProcessGroupStatusSnapshotDTO();
-        selectedNodeSnapshot.setStatusSnapshot(statusDto.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedProcessGroupStatus.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        for (final Map.Entry<NodeIdentifier, ProcessGroupStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final ProcessGroupStatusDTO nodeProcessGroupStatus = entry.getValue();
-            if (nodeProcessGroupStatus == mergedProcessGroupStatus) {
-                continue;
-            }
-
-            final ProcessGroupStatusSnapshotDTO nodeSnapshot = nodeProcessGroupStatus.getAggregateSnapshot();
-            for (final RemoteProcessGroupStatusSnapshotDTO remoteProcessGroupStatus : nodeSnapshot.getRemoteProcessGroupStatusSnapshots()) {
-                final List<String> nodeAuthorizationIssues = remoteProcessGroupStatus.getAuthorizationIssues();
-                if (!nodeAuthorizationIssues.isEmpty()) {
-                    for (final ListIterator<String> iter = nodeAuthorizationIssues.listIterator(); iter.hasNext();) {
-                        final String Issue = iter.next();
-                        iter.set("[" + nodeId.getApiAddress() + ":" + nodeId.getApiPort() + "] -- " + Issue);
-                    }
-                    remoteProcessGroupStatus.setAuthorizationIssues(nodeAuthorizationIssues);
-                }
-            }
-
-            StatusMerger.merge(mergedProcessGroupStatus, nodeProcessGroupStatus, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-
-    private void mergeProcessorStatus(final ProcessorStatusDTO statusDto, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, ProcessorStatusDTO> resultMap) {
-        final ProcessorStatusDTO mergedProcessorStatus = statusDto;
-        mergedProcessorStatus.setNodeSnapshots(new ArrayList<NodeProcessorStatusSnapshotDTO>());
-
-        final NodeProcessorStatusSnapshotDTO selectedNodeSnapshot = new NodeProcessorStatusSnapshotDTO();
-        selectedNodeSnapshot.setStatusSnapshot(statusDto.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedProcessorStatus.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        // merge the other nodes
-        for (final Map.Entry<NodeIdentifier, ProcessorStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final ProcessorStatusDTO nodeProcessorStatus = entry.getValue();
-            if (nodeProcessorStatus == statusDto) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedProcessorStatus, nodeProcessorStatus, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergeConnectionStatus(final ConnectionStatusDTO statusDto, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, ConnectionStatusDTO> resultMap) {
-        final ConnectionStatusDTO mergedConnectionStatus = statusDto;
-        mergedConnectionStatus.setNodeSnapshots(new ArrayList<NodeConnectionStatusSnapshotDTO>());
-
-        final NodeConnectionStatusSnapshotDTO selectedNodeSnapshot = new NodeConnectionStatusSnapshotDTO();
-        selectedNodeSnapshot.setStatusSnapshot(statusDto.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedConnectionStatus.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        // merge the other nodes
-        for (final Map.Entry<NodeIdentifier, ConnectionStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final ConnectionStatusDTO nodeConnectionStatus = entry.getValue();
-            if (nodeConnectionStatus == statusDto) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedConnectionStatus, nodeConnectionStatus, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergePortStatus(final PortStatusDTO statusDto, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, PortStatusDTO> resultMap) {
-        final PortStatusDTO mergedPortStatus = statusDto;
-        mergedPortStatus.setNodeSnapshots(new ArrayList<NodePortStatusSnapshotDTO>());
-
-        final NodePortStatusSnapshotDTO selectedNodeSnapshot = new NodePortStatusSnapshotDTO();
-        selectedNodeSnapshot.setStatusSnapshot(statusDto.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedPortStatus.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        // merge the other nodes
-        for (final Map.Entry<NodeIdentifier, PortStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final PortStatusDTO nodePortStatus = entry.getValue();
-            if (nodePortStatus == statusDto) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedPortStatus, nodePortStatus, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergeRemoteProcessGroupStatus(final RemoteProcessGroupStatusDTO statusDto, final NodeIdentifier selectedNodeId, final Map<NodeIdentifier, RemoteProcessGroupStatusDTO> resultMap) {
-        final RemoteProcessGroupStatusDTO mergedRemoteProcessGroupStatus = statusDto;
-        mergedRemoteProcessGroupStatus.setNodeSnapshots(new ArrayList<NodeRemoteProcessGroupStatusSnapshotDTO>());
-
-        final NodeRemoteProcessGroupStatusSnapshotDTO selectedNodeSnapshot = new NodeRemoteProcessGroupStatusSnapshotDTO();
-        selectedNodeSnapshot.setStatusSnapshot(statusDto.getAggregateSnapshot().clone());
-        selectedNodeSnapshot.setAddress(selectedNodeId.getApiAddress());
-        selectedNodeSnapshot.setApiPort(selectedNodeId.getApiPort());
-        selectedNodeSnapshot.setNodeId(selectedNodeId.getId());
-
-        mergedRemoteProcessGroupStatus.getNodeSnapshots().add(selectedNodeSnapshot);
-
-        // merge the other nodes
-        for (final Map.Entry<NodeIdentifier, RemoteProcessGroupStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final RemoteProcessGroupStatusDTO nodeRemoteProcessGroupStatus = entry.getValue();
-            if (nodeRemoteProcessGroupStatus == statusDto) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedRemoteProcessGroupStatus, nodeRemoteProcessGroupStatus, nodeId.getId(), nodeId.getApiAddress(), nodeId.getApiPort());
-        }
-    }
-
-    private void mergeControllerStatus(final ControllerStatusDTO statusDto, final Map<NodeIdentifier, ControllerStatusDTO> resultMap) {
-        ControllerStatusDTO mergedStatus = statusDto;
-        for (final Map.Entry<NodeIdentifier, ControllerStatusDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final ControllerStatusDTO nodeStatus = entry.getValue();
-
-            final String nodeAddress = nodeId.getApiAddress() + ":" + nodeId.getApiPort();
-            for (final BulletinDTO bulletin : nodeStatus.getBulletins()) {
-                bulletin.setNodeAddress(nodeAddress);
-            }
-            for (final BulletinDTO bulletin : nodeStatus.getControllerServiceBulletins()) {
-                bulletin.setNodeAddress(nodeAddress);
-            }
-            for (final BulletinDTO bulletin : nodeStatus.getReportingTaskBulletins()) {
-                bulletin.setNodeAddress(nodeAddress);
-            }
-
-            if (nodeStatus == mergedStatus) {
-                continue;
-            }
-
-            StatusMerger.merge(mergedStatus, nodeStatus);
-        }
-
-        final int totalNodeCount = getNodeIds().size();
-        final int connectedNodeCount = getNodeIds(Status.CONNECTED).size();
-
-        final List<Bulletin> ncmControllerBulletins = getBulletinRepository().findBulletinsForController();
-        mergedStatus.setBulletins(mergeNCMBulletins(mergedStatus.getBulletins(), ncmControllerBulletins));
-
-        // get the controller service bulletins
-        final BulletinQuery controllerServiceQuery = new BulletinQuery.Builder().sourceType(ComponentType.CONTROLLER_SERVICE).build();
-        final List<Bulletin> ncmServiceBulletins = getBulletinRepository().findBulletins(controllerServiceQuery);
-        mergedStatus.setControllerServiceBulletins(mergeNCMBulletins(mergedStatus.getControllerServiceBulletins(), ncmServiceBulletins));
-
-        // get the reporting task bulletins
-        final BulletinQuery reportingTaskQuery = new BulletinQuery.Builder().sourceType(ComponentType.REPORTING_TASK).build();
-        final List<Bulletin> ncmReportingTaskBulletins = getBulletinRepository().findBulletins(reportingTaskQuery);
-        mergedStatus.setReportingTaskBulletins(mergeNCMBulletins(mergedStatus.getReportingTaskBulletins(), ncmReportingTaskBulletins));
-
-        mergedStatus.setConnectedNodeCount(connectedNodeCount);
-        mergedStatus.setTotalNodeCount(totalNodeCount);
-        StatusMerger.updatePrettyPrintedFields(mergedStatus);
-    }
 
     private List<BulletinDTO> mergeNCMBulletins(final List<BulletinDTO> nodeBulletins, final List<Bulletin> ncmBulletins) {
         if (ncmBulletins == null || ncmBulletins.isEmpty()) {
@@ -2775,33 +2169,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         return mergedBulletins;
     }
 
-    private void mergeBulletinBoard(final BulletinBoardDTO nodeBulletinBoard, final Map<NodeIdentifier, BulletinBoardDTO> resultMap) {
-        final List<BulletinDTO> bulletinDtos = new ArrayList<>();
-        for (final Map.Entry<NodeIdentifier, BulletinBoardDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeId = entry.getKey();
-            final BulletinBoardDTO boardDto = entry.getValue();
-            final String nodeAddress = nodeId.getApiAddress() + ":" + nodeId.getApiPort();
-
-            for (final BulletinDTO bulletin : boardDto.getBulletins()) {
-                bulletin.setNodeAddress(nodeAddress);
-                bulletinDtos.add(bulletin);
-            }
-        }
-
-        Collections.sort(bulletinDtos, new Comparator<BulletinDTO>() {
-            @Override
-            public int compare(final BulletinDTO o1, final BulletinDTO o2) {
-                final int timeComparison = o1.getTimestamp().compareTo(o2.getTimestamp());
-                if (timeComparison != 0) {
-                    return timeComparison;
-                }
-
-                return o1.getNodeAddress().compareTo(o2.getNodeAddress());
-            }
-        });
-
-        nodeBulletinBoard.setBulletins(bulletinDtos);
-    }
 
     /**
      * Creates BulletinDTOs for the specified Bulletins.
@@ -2837,279 +2204,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         return dto;
     }
 
-    private void mergeProvenanceQueryResults(final ProvenanceDTO provenanceDto, final Map<NodeIdentifier, ProvenanceDTO> resultMap, final Set<NodeResponse> problematicResponses) {
-        final ProvenanceResultsDTO results = provenanceDto.getResults();
-        final ProvenanceRequestDTO request = provenanceDto.getRequest();
-        final List<ProvenanceEventDTO> allResults = new ArrayList<>(1024);
-
-        final Set<String> errors = new HashSet<>();
-        Date oldestEventDate = new Date();
-        int percentageComplete = 0;
-        boolean finished = true;
-
-        long totalRecords = 0;
-        for (final Map.Entry<NodeIdentifier, ProvenanceDTO> entry : resultMap.entrySet()) {
-            final NodeIdentifier nodeIdentifier = entry.getKey();
-            final String nodeAddress = nodeIdentifier.getApiAddress() + ":" + nodeIdentifier.getApiPort();
-
-            final ProvenanceDTO nodeDto = entry.getValue();
-            final ProvenanceResultsDTO nodeResultDto = nodeDto.getResults();
-            if (nodeResultDto != null && nodeResultDto.getProvenanceEvents() != null) {
-                // increment the total number of records
-                totalRecords += nodeResultDto.getTotalCount();
-
-                // populate the cluster identifier
-                for (final ProvenanceEventDTO eventDto : nodeResultDto.getProvenanceEvents()) {
-                    eventDto.setClusterNodeId(nodeIdentifier.getId());
-                    eventDto.setClusterNodeAddress(nodeAddress);
-                    // add node identifier to the event's id so that it is unique across cluster
-                    eventDto.setId(nodeIdentifier.getId() + eventDto.getId());
-                    allResults.add(eventDto);
-                }
-            }
-
-            if (nodeResultDto.getOldestEvent() != null && nodeResultDto.getOldestEvent().before(oldestEventDate)) {
-                oldestEventDate = nodeResultDto.getOldestEvent();
-            }
-
-            if (nodeResultDto.getErrors() != null) {
-                for (final String error : nodeResultDto.getErrors()) {
-                    errors.add(nodeAddress + " -- " + error);
-                }
-            }
-
-            percentageComplete += nodeDto.getPercentCompleted();
-            if (!nodeDto.isFinished()) {
-                finished = false;
-            }
-        }
-        percentageComplete /= resultMap.size();
-
-        // consider any problematic responses as errors
-        for (final NodeResponse problematicResponse : problematicResponses) {
-            final NodeIdentifier problemNode = problematicResponse.getNodeId();
-            final String problemNodeAddress = problemNode.getApiAddress() + ":" + problemNode.getApiPort();
-            errors.add(String.format("%s -- Request did not complete successfully (Status code: %s)", problemNodeAddress, problematicResponse.getStatus()));
-        }
-
-        // Since we get back up to the maximum number of results from each node, we need to sort those values and then
-        // grab only the first X number of them. We do a sort based on time, such that the newest are included.
-        // If 2 events have the same timestamp, we do a secondary sort based on Cluster Node Identifier. If those are
-        // equal, we perform a terciary sort based on the the event id
-        Collections.sort(allResults, new Comparator<ProvenanceEventDTO>() {
-            @Override
-            public int compare(final ProvenanceEventDTO o1, final ProvenanceEventDTO o2) {
-                final int eventTimeComparison = o1.getEventTime().compareTo(o2.getEventTime());
-                if (eventTimeComparison != 0) {
-                    return -eventTimeComparison;
-                }
-
-                final String nodeId1 = o1.getClusterNodeId();
-                final String nodeId2 = o2.getClusterNodeId();
-                final int nodeIdComparison;
-                if (nodeId1 == null && nodeId2 == null) {
-                    nodeIdComparison = 0;
-                } else if (nodeId1 == null) {
-                    nodeIdComparison = 1;
-                } else if (nodeId2 == null) {
-                    nodeIdComparison = -1;
-                } else {
-                    nodeIdComparison = -nodeId1.compareTo(nodeId2);
-                }
-
-                if (nodeIdComparison != 0) {
-                    return nodeIdComparison;
-                }
-
-                return -Long.compare(o1.getEventId(), o2.getEventId());
-            }
-        });
-
-        final int maxResults = request.getMaxResults().intValue();
-        final List<ProvenanceEventDTO> selectedResults;
-        if (allResults.size() < maxResults) {
-            selectedResults = allResults;
-        } else {
-            selectedResults = allResults.subList(0, maxResults);
-        }
-
-        // include any errors
-        if (errors.size() > 0) {
-            results.setErrors(errors);
-        }
-
-        results.setTotalCount(totalRecords);
-        results.setTotal(FormatUtils.formatCount(totalRecords));
-        results.setProvenanceEvents(selectedResults);
-        results.setOldestEvent(oldestEventDate);
-        results.setGenerated(new Date());
-        provenanceDto.setPercentCompleted(percentageComplete);
-        provenanceDto.setFinished(finished);
-    }
-
-    private void mergeRemoteProcessGroup(final RemoteProcessGroupDTO remoteProcessGroup, final Map<NodeIdentifier, RemoteProcessGroupDTO> remoteProcessGroupMap) {
-        final RemoteProcessGroupContentsDTO remoteProcessGroupContents = remoteProcessGroup.getContents();
-
-        Boolean mergedIsTargetSecure = null;
-        final List<String> mergedAuthorizationIssues = new ArrayList<>();
-        final Set<RemoteProcessGroupPortDTO> mergedInputPorts = new HashSet<>();
-        final Set<RemoteProcessGroupPortDTO> mergedOutputPorts = new HashSet<>();
-
-        for (final Map.Entry<NodeIdentifier, RemoteProcessGroupDTO> nodeEntry : remoteProcessGroupMap.entrySet()) {
-            final NodeIdentifier nodeId = nodeEntry.getKey();
-            final RemoteProcessGroupDTO nodeRemoteProcessGroupDto = nodeEntry.getValue();
-
-            // merge the  issues
-            final List<String> nodeAuthorizationIssues = nodeRemoteProcessGroupDto.getAuthorizationIssues();
-            if (nodeAuthorizationIssues != null && !nodeAuthorizationIssues.isEmpty()) {
-                for (final String nodeAuthorizationIssue : nodeAuthorizationIssues) {
-                    mergedAuthorizationIssues.add(nodeId.getApiAddress() + ":" + nodeId.getApiPort() + " -- " + nodeAuthorizationIssue);
-                }
-            }
-
-            // use the first target secure flag since they will all be the same
-            final Boolean nodeIsTargetSecure = nodeRemoteProcessGroupDto.isTargetSecure();
-            if (mergedIsTargetSecure == null) {
-                mergedIsTargetSecure = nodeIsTargetSecure;
-            }
-
-            // merge the ports in the contents
-            final RemoteProcessGroupContentsDTO nodeRemoteProcessGroupContentsDto = nodeRemoteProcessGroupDto.getContents();
-            if (remoteProcessGroupContents != null && nodeRemoteProcessGroupContentsDto != null) {
-                if (nodeRemoteProcessGroupContentsDto.getInputPorts() != null) {
-                    mergedInputPorts.addAll(nodeRemoteProcessGroupContentsDto.getInputPorts());
-                }
-                if (nodeRemoteProcessGroupContentsDto.getOutputPorts() != null) {
-                    mergedOutputPorts.addAll(nodeRemoteProcessGroupContentsDto.getOutputPorts());
-                }
-            }
-        }
-
-        if (remoteProcessGroupContents != null) {
-            if (!mergedInputPorts.isEmpty()) {
-                remoteProcessGroupContents.setInputPorts(mergedInputPorts);
-            }
-            if (!mergedOutputPorts.isEmpty()) {
-                remoteProcessGroupContents.setOutputPorts(mergedOutputPorts);
-            }
-        }
-
-        if (mergedIsTargetSecure != null) {
-            remoteProcessGroup.setTargetSecure(mergedIsTargetSecure);
-        }
-
-        if (!mergedAuthorizationIssues.isEmpty()) {
-            remoteProcessGroup.setAuthorizationIssues(mergedAuthorizationIssues);
-        }
-    }
-
-    private void mergeControllerServiceReferences(
-            final Set<ControllerServiceReferencingComponentDTO> referencingComponents, final Map<NodeIdentifier, Set<ControllerServiceReferencingComponentDTO>> referencingComponentMap) {
-        final Map<String, Integer> activeThreadCounts = new HashMap<>();
-        final Map<String, String> states = new HashMap<>();
-        for (final Map.Entry<NodeIdentifier, Set<ControllerServiceReferencingComponentDTO>> nodeEntry : referencingComponentMap.entrySet()) {
-            final Set<ControllerServiceReferencingComponentDTO> nodeReferencingComponents = nodeEntry.getValue();
-
-            // go through all the nodes referencing components
-            if (nodeReferencingComponents != null) {
-                for (final ControllerServiceReferencingComponentDTO nodeReferencingComponent : nodeReferencingComponents) {
-                    // handle active thread counts
-                    if (nodeReferencingComponent.getActiveThreadCount() != null && nodeReferencingComponent.getActiveThreadCount() > 0) {
-                        final Integer current = activeThreadCounts.get(nodeReferencingComponent.getId());
-                        if (current == null) {
-                            activeThreadCounts.put(nodeReferencingComponent.getId(), nodeReferencingComponent.getActiveThreadCount());
-                        } else {
-                            activeThreadCounts.put(nodeReferencingComponent.getId(), nodeReferencingComponent.getActiveThreadCount() + current);
-                        }
-                    }
-
-                    // handle controller service state
-                    final String state = states.get(nodeReferencingComponent.getId());
-                    if (state == null) {
-                        if (ControllerServiceState.DISABLING.name().equals(nodeReferencingComponent.getState())) {
-                            states.put(nodeReferencingComponent.getId(), ControllerServiceState.DISABLING.name());
-                        } else if (ControllerServiceState.ENABLING.name().equals(nodeReferencingComponent.getState())) {
-                            states.put(nodeReferencingComponent.getId(), ControllerServiceState.ENABLING.name());
-                        }
-                    }
-                }
-            }
-        }
-
-        // go through each referencing components
-        for (final ControllerServiceReferencingComponentDTO referencingComponent : referencingComponents) {
-            final Integer activeThreadCount = activeThreadCounts.get(referencingComponent.getId());
-            if (activeThreadCount != null) {
-                referencingComponent.setActiveThreadCount(activeThreadCount);
-            }
-
-            final String state = states.get(referencingComponent.getId());
-            if (state != null) {
-                referencingComponent.setState(state);
-            }
-        }
-    }
-
-    private void mergeControllerService(final ControllerServiceDTO controllerService, final Map<NodeIdentifier, ControllerServiceDTO> controllerServiceMap) {
-        final Map<String, Set<NodeIdentifier>> validationErrorMap = new HashMap<>();
-        final Set<ControllerServiceReferencingComponentDTO> referencingComponents = controllerService.getReferencingComponents();
-        final Map<NodeIdentifier, Set<ControllerServiceReferencingComponentDTO>> nodeReferencingComponentsMap = new HashMap<>();
-
-        String state = null;
-        for (final Map.Entry<NodeIdentifier, ControllerServiceDTO> nodeEntry : controllerServiceMap.entrySet()) {
-            final NodeIdentifier nodeId = nodeEntry.getKey();
-            final ControllerServiceDTO nodeControllerService = nodeEntry.getValue();
-
-            if (state == null) {
-                if (ControllerServiceState.DISABLING.name().equals(nodeControllerService.getState())) {
-                    state = ControllerServiceState.DISABLING.name();
-                } else if (ControllerServiceState.ENABLING.name().equals(nodeControllerService.getState())) {
-                    state = ControllerServiceState.ENABLING.name();
-                }
-            }
-
-            for (final ControllerServiceReferencingComponentDTO nodeReferencingComponents : nodeControllerService.getReferencingComponents()) {
-                nodeReferencingComponentsMap.put(nodeId, nodeReferencingComponents.getReferencingComponents());
-            }
-
-            // merge the validation errors
-            mergeValidationErrors(validationErrorMap, nodeId, nodeControllerService.getValidationErrors());
-        }
-
-        // merge the referencing components
-        mergeControllerServiceReferences(referencingComponents, nodeReferencingComponentsMap);
-
-        // store the 'transition' state is applicable
-        if (state != null) {
-            controllerService.setState(state);
-        }
-
-        // set the merged the validation errors
-        controllerService.setValidationErrors(normalizedMergedValidationErrors(validationErrorMap, controllerServiceMap.size()));
-    }
-
-    private void mergeReportingTask(final ReportingTaskDTO reportingTask, final Map<NodeIdentifier, ReportingTaskDTO> reportingTaskMap) {
-        final Map<String, Set<NodeIdentifier>> validationErrorMap = new HashMap<>();
-
-        int activeThreadCount = 0;
-        for (final Map.Entry<NodeIdentifier, ReportingTaskDTO> nodeEntry : reportingTaskMap.entrySet()) {
-            final NodeIdentifier nodeId = nodeEntry.getKey();
-            final ReportingTaskDTO nodeReportingTask = nodeEntry.getValue();
-
-            if (nodeReportingTask.getActiveThreadCount() != null) {
-                activeThreadCount += nodeReportingTask.getActiveThreadCount();
-            }
-
-            // merge the validation errors
-            mergeValidationErrors(validationErrorMap, nodeId, nodeReportingTask.getValidationErrors());
-        }
-
-        // set the merged active thread counts
-        reportingTask.setActiveThreadCount(activeThreadCount);
-
-        // set the merged the validation errors
-        reportingTask.setValidationErrors(normalizedMergedValidationErrors(validationErrorMap, reportingTaskMap.size()));
-    }
 
     /**
      * Merges the validation errors into the specified map, recording the corresponding node identifier.
@@ -3156,875 +2250,52 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
     }
 
 
-    /**
-     * Merges the listing requests in the specified map into the specified listing request
-     *
-     * @param listingRequest the target listing request
-     * @param listingRequestMap the mapping of all responses being merged
-     */
-    private void mergeListingRequests(final ListingRequestDTO listingRequest, final Map<NodeIdentifier, ListingRequestDTO> listingRequestMap) {
-        final Comparator<FlowFileSummaryDTO> comparator = new Comparator<FlowFileSummaryDTO>() {
-            @Override
-            public int compare(final FlowFileSummaryDTO dto1, final FlowFileSummaryDTO dto2) {
-                int positionCompare = dto1.getPosition().compareTo(dto2.getPosition());
-                if (positionCompare != 0) {
-                    return positionCompare;
-                }
+    // TODO: This is temporary. Only here because while we have NCM, we must merge its bulletins. Once we get rid
+    // of the NCM, this goes away completely.
+    public static final Pattern CONTROLLER_STATUS_URI_PATTERN = Pattern.compile("/nifi-api/controller/status");
 
-                final String address1 = dto1.getClusterNodeAddress();
-                final String address2 = dto2.getClusterNodeAddress();
-                if (address1 == null && address2 == null) {
-                    return 0;
-                }
-                if (address1 == null) {
-                    return 1;
-                }
-                if (address2 == null) {
-                    return -1;
-                }
-                return address1.compareTo(address2);
-            }
-        };
-
-        final NavigableSet<FlowFileSummaryDTO> flowFileSummaries = new TreeSet<>(comparator);
-
-        ListFlowFileState state = null;
-        int numStepsCompleted = 0;
-        int numStepsTotal = 0;
-        int objectCount = 0;
-        long byteCount = 0;
-        boolean finished = true;
-        for (final Map.Entry<NodeIdentifier, ListingRequestDTO> entry : listingRequestMap.entrySet()) {
-            final NodeIdentifier nodeIdentifier = entry.getKey();
-            final String nodeAddress = nodeIdentifier.getApiAddress() + ":" + nodeIdentifier.getApiPort();
-
-            final ListingRequestDTO nodeRequest = entry.getValue();
-
-            numStepsTotal++;
-            if (Boolean.TRUE.equals(nodeRequest.getFinished())) {
-                numStepsCompleted++;
-            }
-
-            final QueueSizeDTO nodeQueueSize = nodeRequest.getQueueSize();
-            objectCount += nodeQueueSize.getObjectCount();
-            byteCount += nodeQueueSize.getByteCount();
-
-            if (!nodeRequest.getFinished()) {
-                finished = false;
-            }
-
-            if (nodeRequest.getLastUpdated().after(listingRequest.getLastUpdated())) {
-                listingRequest.setLastUpdated(nodeRequest.getLastUpdated());
-            }
-
-            // Keep the state with the lowest ordinal value (the "least completed").
-            final ListFlowFileState nodeState = ListFlowFileState.valueOfDescription(nodeRequest.getState());
-            if (state == null || state.compareTo(nodeState) > 0) {
-                state = nodeState;
-            }
-
-            if (nodeRequest.getFlowFileSummaries() != null) {
-                for (final FlowFileSummaryDTO summaryDTO : nodeRequest.getFlowFileSummaries()) {
-                    summaryDTO.setClusterNodeId(nodeIdentifier.getId());
-                    summaryDTO.setClusterNodeAddress(nodeAddress);
-
-                    flowFileSummaries.add(summaryDTO);
-
-                    // Keep the set from growing beyond our max
-                    if (flowFileSummaries.size() > listingRequest.getMaxResults()) {
-                        flowFileSummaries.pollLast();
-                    }
-                }
-            }
-
-            if (nodeRequest.getFailureReason() != null) {
-                listingRequest.setFailureReason(nodeRequest.getFailureReason());
-            }
-        }
-
-        final List<FlowFileSummaryDTO> summaryDTOs = new ArrayList<>(flowFileSummaries);
-        listingRequest.setFlowFileSummaries(summaryDTOs);
-        // depends on invariant if numStepsTotal is 0, so is numStepsCompleted, all steps being completed
-        // would be 1
-        final int percentCompleted = (numStepsTotal == 0) ? 1 : numStepsCompleted / numStepsTotal;
-        listingRequest.setPercentCompleted(percentCompleted);
-        listingRequest.setFinished(finished);
-
-        listingRequest.getQueueSize().setByteCount(byteCount);
-        listingRequest.getQueueSize().setObjectCount(objectCount);
+    private static boolean isControllerStatusEndpoint(final URI uri, final String method) {
+        return "GET".equalsIgnoreCase(method) && CONTROLLER_STATUS_URI_PATTERN.matcher(uri.getPath()).matches();
     }
 
-    /**
-     * Merges the drop requests in the specified map into the specified drop request.
-     *
-     * @param dropRequest the target drop request
-     * @param dropRequestMap the mapping of all responses being merged
-     */
-    private void mergeDropRequests(final DropRequestDTO dropRequest, final Map<NodeIdentifier, DropRequestDTO> dropRequestMap) {
-        boolean nodeWaiting = false;
-        int originalCount = 0;
-        long originalSize = 0;
-        int currentCount = 0;
-        long currentSize = 0;
-        int droppedCount = 0;
-        long droppedSize = 0;
 
-        DropFlowFileState state = null;
-        for (final Map.Entry<NodeIdentifier, DropRequestDTO> nodeEntry : dropRequestMap.entrySet()) {
-            final DropRequestDTO nodeDropRequest = nodeEntry.getValue();
-
-            currentCount += nodeDropRequest.getCurrentCount();
-            currentSize += nodeDropRequest.getCurrentSize();
-            droppedCount += nodeDropRequest.getDroppedCount();
-            droppedSize += nodeDropRequest.getDroppedSize();
-
-            if (nodeDropRequest.getOriginalCount() == null) {
-                nodeWaiting = true;
-            } else {
-                originalCount += nodeDropRequest.getOriginalCount();
-                originalSize += nodeDropRequest.getOriginalSize();
-            }
-
-            final DropFlowFileState nodeState = DropFlowFileState.valueOfDescription(nodeDropRequest.getState());
-            if (state == null || state.compareTo(nodeState) > 0) {
-                state = nodeState;
-            }
-        }
-
-        dropRequest.setCurrentCount(currentCount);
-        dropRequest.setCurrentSize(currentSize);
-        dropRequest.setCurrent(FormatUtils.formatCount(currentCount) + " / " + FormatUtils.formatDataSize(currentSize));
-
-        dropRequest.setDroppedCount(droppedCount);
-        dropRequest.setDroppedSize(droppedSize);
-        dropRequest.setDropped(FormatUtils.formatCount(droppedCount) + " / " + FormatUtils.formatDataSize(droppedSize));
-
-        if (!nodeWaiting) {
-            dropRequest.setOriginalCount(originalCount);
-            dropRequest.setOriginalSize(originalSize);
-            dropRequest.setOriginal(FormatUtils.formatCount(originalCount) + " / " + FormatUtils.formatDataSize(originalSize));
-        }
-
-        if (state != null) {
-            dropRequest.setState(state.toString());
-        }
-    }
 
     // requires write lock to be already acquired unless request is not mutable
     private NodeResponse mergeResponses(final URI uri, final String method, final Set<NodeResponse> nodeResponses, final boolean mutableRequest) {
-        // holds the one response of all the node responses to return to the client
-        NodeResponse clientResponse = null;
-
-        // holds the set of node responses that did not result in a 2XX response
-        final Set<NodeResponse> problematicNodeResponses = new HashSet<>();
-
-        // map updated node to its response
-        final Map<Node, NodeResponse> updatedNodesMap = new HashMap<>();
-        for (final Map.Entry<NodeResponse, Status> entry : httpResponseMapper.map(uri, nodeResponses).entrySet()) {
-
-            final NodeResponse nodeResponse = entry.getKey();
-            final Status nodeStatus = entry.getValue();
-
-            // create new "updated" node by cloning old node and updating status
-            final Node currentNode = getRawNode(nodeResponse.getNodeId().getId());
-            final Node updatedNode = currentNode.clone();
-            updatedNode.setStatus(nodeStatus);
-
-            // map updated node to its response
-            updatedNodesMap.put(updatedNode, nodeResponse);
-
-            // record a client request and any requests that resulted in disconnection
-            if (nodeStatus == Status.CONNECTED) {
-                clientResponse = nodeResponse;
-            } else if (nodeStatus == Status.DISCONNECTED) {
-                problematicNodeResponses.add(nodeResponse);
-            }
-        }
+        final NodeResponse clientResponse = responseMerger.mergeResponses(uri, method, nodeResponses);
 
         // determine if we have at least one response
         final boolean hasClientResponse = clientResponse != null;
         final boolean hasSuccessfulClientResponse = hasClientResponse && clientResponse.is2xx();
 
-        // drain the responses from the socket for those responses not being sent to the client
-        final Set<NodeResponse> nodeResponsesToDrain = new HashSet<>(updatedNodesMap.values());
-        nodeResponsesToDrain.remove(clientResponse);
-
-        if (hasSuccessfulClientResponse && isProcessorEndpoint(uri, method)) {
-            final ProcessorEntity responseEntity = clientResponse.getClientResponse().getEntity(ProcessorEntity.class);
-            final ProcessorDTO processor = responseEntity.getProcessor();
-
-            final Map<NodeIdentifier, ProcessorDTO> processorMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ProcessorEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ProcessorEntity.class);
-                final ProcessorDTO nodeProcessor = nodeResponseEntity.getProcessor();
-                processorMap.put(nodeResponse.getNodeId(), nodeProcessor);
-            }
-
-            mergeProcessorValidationErrors(processor, processorMap);
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProcessorsEndpoint(uri, method)) {
-            final ProcessorsEntity responseEntity = clientResponse.getClientResponse().getEntity(ProcessorsEntity.class);
-            final Set<ProcessorDTO> processors = responseEntity.getProcessors();
-
-            final Map<String, Map<NodeIdentifier, ProcessorDTO>> processorMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ProcessorsEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ProcessorsEntity.class);
-                final Set<ProcessorDTO> nodeProcessors = nodeResponseEntity.getProcessors();
-
-                for (final ProcessorDTO nodeProcessor : nodeProcessors) {
-                    Map<NodeIdentifier, ProcessorDTO> innerMap = processorMap.get(nodeProcessor.getId());
-                    if (innerMap == null) {
-                        innerMap = new HashMap<>();
-                        processorMap.put(nodeProcessor.getId(), innerMap);
-                    }
-
-                    innerMap.put(nodeResponse.getNodeId(), nodeProcessor);
-                }
-            }
-
-            for (final ProcessorDTO processor : processors) {
-                final String procId = processor.getId();
-                final Map<NodeIdentifier, ProcessorDTO> mergeMap = processorMap.get(procId);
-
-                mergeProcessorValidationErrors(processor, mergeMap);
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProcessGroupEndpoint(uri, method)) {
-            final ProcessGroupEntity responseEntity = clientResponse.getClientResponse().getEntity(ProcessGroupEntity.class);
-            final ProcessGroupDTO responseDto = responseEntity.getProcessGroup();
-
-            final FlowSnippetDTO contents = responseDto.getContents();
-            if (contents == null) {
-                if (!nodeResponsesToDrain.isEmpty()) {
-                    drainResponses(nodeResponsesToDrain);
-                }
-            } else {
-                final Map<String, Map<NodeIdentifier, ProcessorDTO>> processorMap = new HashMap<>();
-                final Map<String, Map<NodeIdentifier, RemoteProcessGroupDTO>> remoteProcessGroupMap = new HashMap<>();
-
-                for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                    if (problematicNodeResponses.contains(nodeResponse)) {
-                        continue;
-                    }
-
-                    final ProcessGroupEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ProcessGroupEntity.class);
-                    final ProcessGroupDTO nodeProcessGroup = nodeResponseEntity.getProcessGroup();
-
-                    for (final ProcessorDTO nodeProcessor : nodeProcessGroup.getContents().getProcessors()) {
-                        Map<NodeIdentifier, ProcessorDTO> innerMap = processorMap.get(nodeProcessor.getId());
-                        if (innerMap == null) {
-                            innerMap = new HashMap<>();
-                            processorMap.put(nodeProcessor.getId(), innerMap);
-                        }
-
-                        innerMap.put(nodeResponse.getNodeId(), nodeProcessor);
-                    }
-
-                    for (final RemoteProcessGroupDTO nodeRemoteProcessGroup : nodeProcessGroup.getContents().getRemoteProcessGroups()) {
-                        Map<NodeIdentifier, RemoteProcessGroupDTO> innerMap = remoteProcessGroupMap.get(nodeRemoteProcessGroup.getId());
-                        if (innerMap == null) {
-                            innerMap = new HashMap<>();
-                            remoteProcessGroupMap.put(nodeRemoteProcessGroup.getId(), innerMap);
-                        }
-
-                        innerMap.put(nodeResponse.getNodeId(), nodeRemoteProcessGroup);
-                    }
-                }
-
-                for (final ProcessorDTO processor : contents.getProcessors()) {
-                    final String procId = processor.getId();
-                    final Map<NodeIdentifier, ProcessorDTO> mergeMap = processorMap.get(procId);
-
-                    mergeProcessorValidationErrors(processor, mergeMap);
-                }
-
-                for (final RemoteProcessGroupDTO remoteProcessGroup : contents.getRemoteProcessGroups()) {
-                    if (remoteProcessGroup.getContents() != null) {
-                        final String remoteProcessGroupId = remoteProcessGroup.getId();
-                        final Map<NodeIdentifier, RemoteProcessGroupDTO> mergeMap = remoteProcessGroupMap.get(remoteProcessGroupId);
-
-                        mergeRemoteProcessGroup(remoteProcessGroup, mergeMap);
-                    }
-                }
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && (isTemplateEndpoint(uri, method) || isFlowSnippetEndpoint(uri, method))) {
-            final FlowSnippetEntity responseEntity = clientResponse.getClientResponse().getEntity(FlowSnippetEntity.class);
-            final FlowSnippetDTO contents = responseEntity.getContents();
-
-            if (contents == null) {
-                if (!nodeResponsesToDrain.isEmpty()) {
-                    drainResponses(nodeResponsesToDrain);
-                }
-            } else {
-                final Map<String, Map<NodeIdentifier, ProcessorDTO>> processorMap = new HashMap<>();
-                final Map<String, Map<NodeIdentifier, RemoteProcessGroupDTO>> remoteProcessGroupMap = new HashMap<>();
-
-                for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                    if (problematicNodeResponses.contains(nodeResponse)) {
-                        continue;
-                    }
-
-                    final FlowSnippetEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(FlowSnippetEntity.class);
-                    final FlowSnippetDTO nodeContents = nodeResponseEntity.getContents();
-
-                    for (final ProcessorDTO nodeProcessor : nodeContents.getProcessors()) {
-                        Map<NodeIdentifier, ProcessorDTO> innerMap = processorMap.get(nodeProcessor.getId());
-                        if (innerMap == null) {
-                            innerMap = new HashMap<>();
-                            processorMap.put(nodeProcessor.getId(), innerMap);
-                        }
-
-                        innerMap.put(nodeResponse.getNodeId(), nodeProcessor);
-                    }
-
-                    for (final RemoteProcessGroupDTO nodeRemoteProcessGroup : nodeContents.getRemoteProcessGroups()) {
-                        Map<NodeIdentifier, RemoteProcessGroupDTO> innerMap = remoteProcessGroupMap.get(nodeRemoteProcessGroup.getId());
-                        if (innerMap == null) {
-                            innerMap = new HashMap<>();
-                            remoteProcessGroupMap.put(nodeRemoteProcessGroup.getId(), innerMap);
-                        }
-
-                        innerMap.put(nodeResponse.getNodeId(), nodeRemoteProcessGroup);
-                    }
-                }
-
-                for (final ProcessorDTO processor : contents.getProcessors()) {
-                    final String procId = processor.getId();
-                    final Map<NodeIdentifier, ProcessorDTO> mergeMap = processorMap.get(procId);
-
-                    mergeProcessorValidationErrors(processor, mergeMap);
-                }
-
-                for (final RemoteProcessGroupDTO remoteProcessGroup : contents.getRemoteProcessGroups()) {
-                    if (remoteProcessGroup.getContents() != null) {
-                        final String remoteProcessGroupId = remoteProcessGroup.getId();
-                        final Map<NodeIdentifier, RemoteProcessGroupDTO> mergeMap = remoteProcessGroupMap.get(remoteProcessGroupId);
-
-                        mergeRemoteProcessGroup(remoteProcessGroup, mergeMap);
-                    }
-                }
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isRemoteProcessGroupEndpoint(uri, method)) {
-            final RemoteProcessGroupEntity responseEntity = clientResponse.getClientResponse().getEntity(RemoteProcessGroupEntity.class);
-            final RemoteProcessGroupDTO remoteProcessGroup = responseEntity.getRemoteProcessGroup();
-
-            final Map<NodeIdentifier, RemoteProcessGroupDTO> remoteProcessGroupMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final RemoteProcessGroupEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(RemoteProcessGroupEntity.class);
-                final RemoteProcessGroupDTO nodeRemoteProcessGroup = nodeResponseEntity.getRemoteProcessGroup();
-
-                remoteProcessGroupMap.put(nodeResponse.getNodeId(), nodeRemoteProcessGroup);
-            }
-            mergeRemoteProcessGroup(remoteProcessGroup, remoteProcessGroupMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isRemoteProcessGroupsEndpoint(uri, method)) {
-            final RemoteProcessGroupsEntity responseEntity = clientResponse.getClientResponse().getEntity(RemoteProcessGroupsEntity.class);
-            final Set<RemoteProcessGroupDTO> remoteProcessGroups = responseEntity.getRemoteProcessGroups();
-
-            final Map<String, Map<NodeIdentifier, RemoteProcessGroupDTO>> remoteProcessGroupMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final RemoteProcessGroupsEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(RemoteProcessGroupsEntity.class);
-                final Set<RemoteProcessGroupDTO> nodeRemoteProcessGroups = nodeResponseEntity.getRemoteProcessGroups();
-
-                for (final RemoteProcessGroupDTO nodeRemoteProcessGroup : nodeRemoteProcessGroups) {
-                    Map<NodeIdentifier, RemoteProcessGroupDTO> innerMap = remoteProcessGroupMap.get(nodeRemoteProcessGroup.getId());
-                    if (innerMap == null) {
-                        innerMap = new HashMap<>();
-                        remoteProcessGroupMap.put(nodeRemoteProcessGroup.getId(), innerMap);
-                    }
-
-                    innerMap.put(nodeResponse.getNodeId(), nodeRemoteProcessGroup);
-                }
-            }
-
-            for (final RemoteProcessGroupDTO remoteProcessGroup : remoteProcessGroups) {
-                final String remoteProcessGroupId = remoteProcessGroup.getId();
-                final Map<NodeIdentifier, RemoteProcessGroupDTO> mergeMap = remoteProcessGroupMap.get(remoteProcessGroupId);
-
-                mergeRemoteProcessGroup(remoteProcessGroup, mergeMap);
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProvenanceQueryEndpoint(uri, method)) {
-            final ProvenanceEntity responseEntity = clientResponse.getClientResponse().getEntity(ProvenanceEntity.class);
-            final ProvenanceDTO query = responseEntity.getProvenance();
-
-            final Map<NodeIdentifier, ProvenanceDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ProvenanceEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ProvenanceEntity.class);
-                final ProvenanceDTO nodeQuery = nodeResponseEntity.getProvenance();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeQuery);
-            }
-            mergeProvenanceQueryResults(query, resultsMap, problematicNodeResponses);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProvenanceEventEndpoint(uri, method)) {
-            final ProvenanceEventEntity responseEntity = clientResponse.getClientResponse().getEntity(ProvenanceEventEntity.class);
-            final ProvenanceEventDTO event = responseEntity.getProvenanceEvent();
-
-            // this request was sent to a specific node... populate its details
-            final NodeIdentifier nodeId = clientResponse.getNodeId();
-            event.setClusterNodeId(nodeId.getId());
-            event.setClusterNodeAddress(nodeId.getApiAddress() + ":" + nodeId.getApiPort());
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isControllerServiceEndpoint(uri, method)) {
-            final ControllerServiceEntity responseEntity = clientResponse.getClientResponse().getEntity(ControllerServiceEntity.class);
-            final ControllerServiceDTO controllerService = responseEntity.getControllerService();
-
-            final Map<NodeIdentifier, ControllerServiceDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ControllerServiceEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ControllerServiceEntity.class);
-                final ControllerServiceDTO nodeControllerService = nodeResponseEntity.getControllerService();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeControllerService);
-            }
-            mergeControllerService(controllerService, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isControllerServicesEndpoint(uri, method)) {
-            final ControllerServicesEntity responseEntity = clientResponse.getClientResponse().getEntity(ControllerServicesEntity.class);
-            final Set<ControllerServiceDTO> controllerServices = responseEntity.getControllerServices();
-
-            final Map<String, Map<NodeIdentifier, ControllerServiceDTO>> controllerServiceMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ControllerServicesEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ControllerServicesEntity.class);
-                final Set<ControllerServiceDTO> nodeControllerServices = nodeResponseEntity.getControllerServices();
-
-                for (final ControllerServiceDTO nodeControllerService : nodeControllerServices) {
-                    Map<NodeIdentifier, ControllerServiceDTO> innerMap = controllerServiceMap.get(nodeControllerService.getId());
-                    if (innerMap == null) {
-                        innerMap = new HashMap<>();
-                        controllerServiceMap.put(nodeControllerService.getId(), innerMap);
-                    }
-
-                    innerMap.put(nodeResponse.getNodeId(), nodeControllerService);
-                }
-            }
-
-            for (final ControllerServiceDTO controllerService : controllerServices) {
-                final String procId = controllerService.getId();
-                final Map<NodeIdentifier, ControllerServiceDTO> mergeMap = controllerServiceMap.get(procId);
-
-                mergeControllerService(controllerService, mergeMap);
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isControllerServiceReferenceEndpoint(uri, method)) {
-            final ControllerServiceReferencingComponentsEntity responseEntity = clientResponse.getClientResponse().getEntity(ControllerServiceReferencingComponentsEntity.class);
-            final Set<ControllerServiceReferencingComponentDTO> referencingComponents = responseEntity.getControllerServiceReferencingComponents();
-
-            final Map<NodeIdentifier, Set<ControllerServiceReferencingComponentDTO>> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ControllerServiceReferencingComponentsEntity nodeResponseEntity
-                        = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ControllerServiceReferencingComponentsEntity.class);
-                final Set<ControllerServiceReferencingComponentDTO> nodeReferencingComponents = nodeResponseEntity.getControllerServiceReferencingComponents();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeReferencingComponents);
-            }
-            mergeControllerServiceReferences(referencingComponents, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isReportingTaskEndpoint(uri, method)) {
-            final ReportingTaskEntity responseEntity = clientResponse.getClientResponse().getEntity(ReportingTaskEntity.class);
-            final ReportingTaskDTO reportingTask = responseEntity.getReportingTask();
-
-            final Map<NodeIdentifier, ReportingTaskDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ReportingTaskEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ReportingTaskEntity.class);
-                final ReportingTaskDTO nodeReportingTask = nodeResponseEntity.getReportingTask();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeReportingTask);
-            }
-            mergeReportingTask(reportingTask, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isReportingTasksEndpoint(uri, method)) {
-            final ReportingTasksEntity responseEntity = clientResponse.getClientResponse().getEntity(ReportingTasksEntity.class);
-            final Set<ReportingTaskDTO> reportingTaskSet = responseEntity.getReportingTasks();
-
-            final Map<String, Map<NodeIdentifier, ReportingTaskDTO>> reportingTaskMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ReportingTasksEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ReportingTasksEntity.class);
-                final Set<ReportingTaskDTO> nodeReportingTasks = nodeResponseEntity.getReportingTasks();
-
-                for (final ReportingTaskDTO nodeReportingTask : nodeReportingTasks) {
-                    Map<NodeIdentifier, ReportingTaskDTO> innerMap = reportingTaskMap.get(nodeReportingTask.getId());
-                    if (innerMap == null) {
-                        innerMap = new HashMap<>();
-                        reportingTaskMap.put(nodeReportingTask.getId(), innerMap);
-                    }
-
-                    innerMap.put(nodeResponse.getNodeId(), nodeReportingTask);
-                }
-            }
-
-            for (final ReportingTaskDTO reportingTask : reportingTaskSet) {
-                final String procId = reportingTask.getId();
-                final Map<NodeIdentifier, ReportingTaskDTO> mergeMap = reportingTaskMap.get(procId);
-
-                mergeReportingTask(reportingTask, mergeMap);
-            }
-
-            // create a new client response
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isDropRequestEndpoint(uri, method)) {
-            final DropRequestEntity responseEntity = clientResponse.getClientResponse().getEntity(DropRequestEntity.class);
-            final DropRequestDTO dropRequest = responseEntity.getDropRequest();
-
-            final Map<NodeIdentifier, DropRequestDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final DropRequestEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(DropRequestEntity.class);
-                final DropRequestDTO nodeDropRequest = nodeResponseEntity.getDropRequest();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeDropRequest);
-            }
-            mergeDropRequests(dropRequest, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isListFlowFilesEndpoint(uri, method)) {
-            final ListingRequestEntity responseEntity = clientResponse.getClientResponse().getEntity(ListingRequestEntity.class);
-            final ListingRequestDTO listingRequest = responseEntity.getListingRequest();
-
-            final Map<NodeIdentifier, ListingRequestDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ListingRequestEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ListingRequestEntity.class);
-                final ListingRequestDTO nodeListingRequest = nodeResponseEntity.getListingRequest();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeListingRequest);
-            }
-            mergeListingRequests(listingRequest, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && (isProcessorStateEndpoint(uri, method) || isControllerServiceStateEndpoint(uri, method) || isReportingTaskStateEndpoint(uri, method))) {
-            final ComponentStateEntity responseEntity = clientResponse.getClientResponse().getEntity(ComponentStateEntity.class);
-            final ComponentStateDTO componentState = responseEntity.getComponentState();
-
-            final Map<NodeIdentifier, ComponentStateDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ComponentStateEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ComponentStateEntity.class);
-                final ComponentStateDTO nodeComponentState = nodeResponseEntity.getComponentState();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeComponentState);
-            }
-            mergeComponentState(componentState, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isGroupStatusEndpoint(uri, method)) {
-            final ProcessGroupStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(ProcessGroupStatusEntity.class);
-            final ProcessGroupStatusDTO statusRequest = responseEntity.getProcessGroupStatus();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, ProcessGroupStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ProcessGroupStatusEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(ProcessGroupStatusEntity.class);
-                }
-
-                final ProcessGroupStatusDTO nodeStatus = nodeResponseEntity.getProcessGroupStatus();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeGroupStatus(statusRequest, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProcessorStatusEndpoint(uri, method)) {
-            final ProcessorStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(ProcessorStatusEntity.class);
-            final ProcessorStatusDTO statusRequest = responseEntity.getProcessorStatus();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, ProcessorStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ProcessorStatusEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(ProcessorStatusEntity.class);
-                }
-
-                final ProcessorStatusDTO nodeStatus = nodeResponseEntity.getProcessorStatus();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeProcessorStatus(statusRequest, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isConnectionStatusEndpoint(uri, method)) {
-            final ConnectionStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(ConnectionStatusEntity.class);
-            final ConnectionStatusDTO statusRequest = responseEntity.getConnectionStatus();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, ConnectionStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ConnectionStatusEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(ConnectionStatusEntity.class);
-                }
-
-                final ConnectionStatusDTO nodeStatus = nodeResponseEntity.getConnectionStatus();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeConnectionStatus(statusRequest, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && (isInputPortStatusEndpoint(uri, method) || isOutputPortStatusEndpoint(uri, method))) {
-            final PortStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(PortStatusEntity.class);
-            final PortStatusDTO statusRequest = responseEntity.getPortStatus();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, PortStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final PortStatusEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(PortStatusEntity.class);
-                }
-
-                final PortStatusDTO nodeStatus = nodeResponseEntity.getPortStatus();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergePortStatus(statusRequest, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isRemoteProcessGroupStatusEndpoint(uri, method)) {
-            final RemoteProcessGroupStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(RemoteProcessGroupStatusEntity.class);
-            final RemoteProcessGroupStatusDTO statusRequest = responseEntity.getRemoteProcessGroupStatus();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, RemoteProcessGroupStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final RemoteProcessGroupStatusEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(RemoteProcessGroupStatusEntity.class);
-                }
-
-                final RemoteProcessGroupStatusDTO nodeStatus = nodeResponseEntity.getRemoteProcessGroupStatus();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeRemoteProcessGroupStatus(statusRequest, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isControllerStatusEndpoint(uri, method)) {
-            final ControllerStatusEntity responseEntity = clientResponse.getClientResponse().getEntity(ControllerStatusEntity.class);
-            final ControllerStatusDTO statusRequest = responseEntity.getControllerStatus();
-
-            final Map<NodeIdentifier, ControllerStatusDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final ControllerStatusEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(ControllerStatusEntity.class);
-                final ControllerStatusDTO nodeStatus = nodeResponseEntity.getControllerStatus();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeControllerStatus(statusRequest, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isBulletinBoardEndpoint(uri, method)) {
-            final BulletinBoardEntity responseEntity = clientResponse.getClientResponse().getEntity(BulletinBoardEntity.class);
-            final BulletinBoardDTO responseDto = responseEntity.getBulletinBoard();
-
-            final Map<NodeIdentifier, BulletinBoardDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final BulletinBoardEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(BulletinBoardEntity.class);
-                final BulletinBoardDTO nodeStatus = nodeResponseEntity.getBulletinBoard();
-
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeBulletinBoard(responseDto, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isProcessorStatusHistoryEndpoint(uri, method)) {
-            final Map<String, MetricDescriptor<?>> metricDescriptors = new HashMap<>();
-            for (final ProcessorStatusDescriptor descriptor : ProcessorStatusDescriptor.values()) {
-                metricDescriptors.put(descriptor.getField(), descriptor.getDescriptor());
-            }
-
-            clientResponse = mergeStatusHistoryResponses(clientResponse, updatedNodesMap, problematicNodeResponses, metricDescriptors);
-        } else if (hasSuccessfulClientResponse && isConnectionStatusHistoryEndpoint(uri, method)) {
-            final Map<String, MetricDescriptor<?>> metricDescriptors = new HashMap<>();
-            for (final ConnectionStatusDescriptor descriptor : ConnectionStatusDescriptor.values()) {
-                metricDescriptors.put(descriptor.getField(), descriptor.getDescriptor());
-            }
-
-            clientResponse = mergeStatusHistoryResponses(clientResponse, updatedNodesMap, problematicNodeResponses, metricDescriptors);
-        } else if (hasSuccessfulClientResponse && isProcessGroupStatusHistoryEndpoint(uri, method)) {
-            final Map<String, MetricDescriptor<?>> metricDescriptors = new HashMap<>();
-            for (final ProcessGroupStatusDescriptor descriptor : ProcessGroupStatusDescriptor.values()) {
-                metricDescriptors.put(descriptor.getField(), descriptor.getDescriptor());
-            }
-
-            clientResponse = mergeStatusHistoryResponses(clientResponse, updatedNodesMap, problematicNodeResponses, metricDescriptors);
-        } else if (hasSuccessfulClientResponse && isRemoteProcessGroupStatusHistoryEndpoint(uri, method)) {
-            final Map<String, MetricDescriptor<?>> metricDescriptors = new HashMap<>();
-            for (final RemoteProcessGroupStatusDescriptor descriptor : RemoteProcessGroupStatusDescriptor.values()) {
-                metricDescriptors.put(descriptor.getField(), descriptor.getDescriptor());
-            }
-
-            clientResponse = mergeStatusHistoryResponses(clientResponse, updatedNodesMap, problematicNodeResponses, metricDescriptors);
-        } else if (hasSuccessfulClientResponse && isSystemDiagnosticsEndpoint(uri, method)) {
-            final SystemDiagnosticsEntity responseEntity = clientResponse.getClientResponse().getEntity(SystemDiagnosticsEntity.class);
-            final SystemDiagnosticsDTO responseDto = responseEntity.getSystemDiagnostics();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, SystemDiagnosticsDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final SystemDiagnosticsEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(SystemDiagnosticsEntity.class);
-                }
-
-                final SystemDiagnosticsDTO nodeStatus = nodeResponseEntity.getSystemDiagnostics();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeSystemDiagnostics(responseDto, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else if (hasSuccessfulClientResponse && isCountersEndpoint(uri, method)) {
-            final CountersEntity responseEntity = clientResponse.getClientResponse().getEntity(CountersEntity.class);
-            final CountersDTO responseDto = responseEntity.getCounters();
-
-            NodeIdentifier nodeIdentifier = null;
-
-            final Map<NodeIdentifier, CountersDTO> resultsMap = new HashMap<>();
-            for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-                if (problematicNodeResponses.contains(nodeResponse)) {
-                    continue;
-                }
-
-                final CountersEntity nodeResponseEntity;
-                if (nodeResponse == clientResponse) {
-                    nodeIdentifier = nodeResponse.getNodeId();
-                    nodeResponseEntity = responseEntity;
-                } else {
-                    nodeResponseEntity = nodeResponse.getClientResponse().getEntity(CountersEntity.class);
-                }
-
-                final CountersDTO nodeStatus = nodeResponseEntity.getCounters();
-                resultsMap.put(nodeResponse.getNodeId(), nodeStatus);
-            }
-            mergeCounters(responseDto, nodeIdentifier, resultsMap);
-
-            clientResponse = new NodeResponse(clientResponse, responseEntity);
-        } else {
-            if (!nodeResponsesToDrain.isEmpty()) {
-                drainResponses(nodeResponsesToDrain);
-            }
+        if (hasSuccessfulClientResponse && isControllerStatusEndpoint(uri, method)) {
+
+            // for now, we need to merge the NCM's bulletins too.
+            // TODO: Remove this logic when we can.
+            final ControllerStatusEntity responseEntity = (ControllerStatusEntity) clientResponse.getUpdatedEntity();
+            final ControllerStatusDTO mergedStatus = responseEntity.getControllerStatus();
+
+            final int totalNodeCount = getNodeIds().size();
+            final int connectedNodeCount = getNodeIds(Status.CONNECTED).size();
+
+            final List<Bulletin> ncmControllerBulletins = getBulletinRepository().findBulletinsForController();
+            mergedStatus.setBulletins(mergeNCMBulletins(mergedStatus.getBulletins(), ncmControllerBulletins));
+
+            // get the controller service bulletins
+            final BulletinQuery controllerServiceQuery = new BulletinQuery.Builder().sourceType(ComponentType.CONTROLLER_SERVICE).build();
+            final List<Bulletin> ncmServiceBulletins = getBulletinRepository().findBulletins(controllerServiceQuery);
+            mergedStatus.setControllerServiceBulletins(mergeNCMBulletins(mergedStatus.getControllerServiceBulletins(), ncmServiceBulletins));
+
+            // get the reporting task bulletins
+            final BulletinQuery reportingTaskQuery = new BulletinQuery.Builder().sourceType(ComponentType.REPORTING_TASK).build();
+            final List<Bulletin> ncmReportingTaskBulletins = getBulletinRepository().findBulletins(reportingTaskQuery);
+            mergedStatus.setReportingTaskBulletins(mergeNCMBulletins(mergedStatus.getReportingTaskBulletins(), ncmReportingTaskBulletins));
+
+            mergedStatus.setConnectedNodeCount(connectedNodeCount);
+            mergedStatus.setTotalNodeCount(totalNodeCount);
+            StatusMerger.updatePrettyPrintedFields(mergedStatus);
         }
+
 
         /*
          * Nodes that encountered issues handling the request are marked as
@@ -4033,6 +2304,7 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
          * state even if they had problems handling the request.
          */
         if (mutableRequest) {
+            final Set<NodeResponse> problematicNodeResponses = responseMerger.getProblematicNodeResponses(nodeResponses);
 
             // all nodes failed
             final boolean allNodesFailed = problematicNodeResponses.size() == nodeResponses.size();
@@ -4043,9 +2315,8 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
 
             // ensure nodes stay connected in certain scenarios
             if (allNodesFailed || someNodesFailedMissingCounter) {
-                for (final Map.Entry<Node, NodeResponse> updatedNodeEntry : updatedNodesMap.entrySet()) {
-                    final NodeResponse nodeResponse = updatedNodeEntry.getValue();
-                    final Node node = updatedNodeEntry.getKey();
+                for (final NodeResponse nodeResponse : nodeResponses) {
+                    final Node node = getRawNode(nodeResponse.getNodeId().getId());
 
                     if (problematicNodeResponses.contains(nodeResponse)) {
                         node.setStatus(Status.CONNECTED);
@@ -4053,10 +2324,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
                     }
                 }
             }
-
-            // set the updated nodes
-            nodes.removeAll(updatedNodesMap.keySet());
-            nodes.addAll(updatedNodesMap.keySet());
 
             // notify service of updated node set
             notifyDataFlowManagementServiceOfNodeStatusChange();
@@ -4076,47 +2343,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         }
 
         return clientResponse;
-    }
-
-
-    private NodeResponse mergeStatusHistoryResponses(NodeResponse clientResponse, Map<Node, NodeResponse> updatedNodesMap, Set<NodeResponse> problematicNodeResponses,
-        Map<String, MetricDescriptor<?>> metricDescriptors) {
-        final StatusHistoryEntity responseEntity = clientResponse.getClientResponse().getEntity(StatusHistoryEntity.class);
-
-        StatusHistoryDTO lastStatusHistory = null;
-        final List<NodeStatusSnapshotsDTO> nodeStatusSnapshots = new ArrayList<>(updatedNodesMap.size());
-        for (final NodeResponse nodeResponse : updatedNodesMap.values()) {
-            if (problematicNodeResponses.contains(nodeResponse)) {
-                continue;
-            }
-
-            final StatusHistoryEntity nodeResponseEntity = nodeResponse == clientResponse ? responseEntity : nodeResponse.getClientResponse().getEntity(StatusHistoryEntity.class);
-            final StatusHistoryDTO nodeStatus = nodeResponseEntity.getStatusHistory();
-            lastStatusHistory = nodeStatus;
-
-            final NodeIdentifier nodeId = nodeResponse.getNodeId();
-            final NodeStatusSnapshotsDTO nodeStatusSnapshot = new NodeStatusSnapshotsDTO();
-            nodeStatusSnapshot.setNodeId(nodeId.getId());
-            nodeStatusSnapshot.setAddress(nodeId.getApiAddress());
-            nodeStatusSnapshot.setApiPort(nodeId.getApiPort());
-            nodeStatusSnapshot.setStatusSnapshots(nodeStatus.getAggregateSnapshots());
-            nodeStatusSnapshots.add(nodeStatusSnapshot);
-        }
-
-        final StatusHistoryDTO clusterStatusHistory = new StatusHistoryDTO();
-        clusterStatusHistory.setAggregateSnapshots(mergeStatusHistories(nodeStatusSnapshots, metricDescriptors));
-        clusterStatusHistory.setGenerated(new Date());
-        clusterStatusHistory.setNodeSnapshots(nodeStatusSnapshots);
-        if (lastStatusHistory != null) {
-            clusterStatusHistory.setComponentDetails(lastStatusHistory.getComponentDetails());
-            clusterStatusHistory.setFieldDescriptors(lastStatusHistory.getFieldDescriptors());
-        }
-
-        final StatusHistoryEntity clusterEntity = new StatusHistoryEntity();
-        clusterEntity.setStatusHistory(clusterStatusHistory);
-        clusterEntity.setRevision(responseEntity.getRevision());
-
-        return new NodeResponse(clientResponse, clusterEntity);
     }
 
 
@@ -4143,42 +2369,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
         return false;
     }
 
-    /**
-     * Drains the node responses off of the socket to ensure that the socket is appropriately cleaned-up.
-     *
-     * @param nodeResponses the collection of node responses
-     */
-    private void drainResponses(final Collection<NodeResponse> nodeResponses) {
-        // fail fast if nothing to do
-        if (nodeResponses.isEmpty()) {
-            return;
-        }
-
-        final ExecutorService executorService = Executors.newFixedThreadPool(properties.getClusterManagerProtocolThreads());
-        final CompletionService<Void> completionService = new ExecutorCompletionService<>(executorService);
-        for (final NodeResponse nodeResponse : nodeResponses) {
-            // if we received a response, then clear out the response data
-            if (!nodeResponse.hasThrowable()) {
-                completionService.submit(new Runnable() {
-                    @Override
-                    public void run() {
-                        try {
-                            try (final OutputStream drain = new OutputStream() {
-                                @Override
-                                public void write(final int b) { /* drain response */ }
-                            }) {
-                                ((StreamingOutput) nodeResponse.getResponse().getEntity()).write(drain);
-                            }
-                        } catch (final IOException | WebApplicationException ex) {
-                            logger.info("Failed clearing out non-client response buffer due to: " + ex, ex);
-                        }
-                    }
-                }, null);
-            }
-        }
-
-        executorService.shutdown();
-    }
 
     /**
      * A helper method to disconnect nodes that returned unsuccessful HTTP responses because of a replicated request. Disconnection requests are sent concurrently.
@@ -4482,88 +2672,6 @@ public class WebClusterManager implements HttpClusterManager, ProtocolHandler, C
     @Override
     public BulletinRepository getBulletinRepository() {
         return bulletinRepository;
-    }
-
-
-    public static Date normalizeStatusSnapshotDate(final Date toNormalize, final long numMillis) {
-        final long time = toNormalize.getTime();
-        return new Date(time - time % numMillis);
-    }
-
-
-    private List<StatusSnapshotDTO> aggregate(Map<Date, List<StatusSnapshot>> snapshotsToAggregate) {
-        // Aggregate the snapshots
-        final List<StatusSnapshotDTO> aggregatedSnapshotDtos = new ArrayList<>();
-        for (final Map.Entry<Date, List<StatusSnapshot>> entry : snapshotsToAggregate.entrySet()) {
-            final List<StatusSnapshot> snapshots = entry.getValue();
-            final StatusSnapshot reducedSnapshot = snapshots.get(0).getValueReducer().reduce(snapshots);
-
-            final StatusSnapshotDTO dto = new StatusSnapshotDTO();
-            dto.setTimestamp(reducedSnapshot.getTimestamp());
-            dto.setStatusMetrics(StatusHistoryUtil.createStatusSnapshotDto(reducedSnapshot).getStatusMetrics());
-
-            aggregatedSnapshotDtos.add(dto);
-        }
-
-        return aggregatedSnapshotDtos;
-    }
-
-
-    private StatusSnapshot createSnapshot(final StatusSnapshotDTO snapshotDto, final Map<String, MetricDescriptor<?>> metricDescriptors) {
-        final StandardStatusSnapshot snapshot = new StandardStatusSnapshot();
-        snapshot.setTimestamp(snapshotDto.getTimestamp());
-
-        final Map<String, Long> metrics = snapshotDto.getStatusMetrics();
-        for (final Map.Entry<String, Long> entry : metrics.entrySet()) {
-            final String metricId = entry.getKey();
-            final Long value = entry.getValue();
-
-            final MetricDescriptor<?> descriptor = metricDescriptors.get(metricId);
-            if (descriptor != null) {
-                snapshot.addStatusMetric(descriptor, value);
-            }
-        }
-
-        return snapshot;
-    }
-
-    private List<StatusSnapshotDTO> mergeStatusHistories(final List<NodeStatusSnapshotsDTO> nodeStatusSnapshots, final Map<String, MetricDescriptor<?>> metricDescriptors) {
-        // We want a Map<Date, List<StatusSnapshot>>, which is a Map of "normalized Date" (i.e., a time range, essentially)
-        // to all Snapshots for that time. The list will contain one snapshot for each node. However, we can have the case
-        // where the NCM has a different value for the componentStatusSnapshotMillis than the nodes have. In this case,
-        // we end up with multiple entries in the List<StatusSnapshot> for the same node/timestamp, which skews our aggregate
-        // results. In order to avoid this, we will use only the latest snapshot for a node that falls into the the time range
-        // of interest.
-        // To accomplish this, we have an intermediate data structure, which is a Map of "normalized Date" to an inner Map
-        // of Node Identifier to StatusSnapshot. We then will flatten this Map and aggregate the results.
-        final Map<Date, Map<String, StatusSnapshot>> dateToNodeSnapshots = new TreeMap<>();
-
-        // group status snapshot's for each node by date
-        for (final NodeStatusSnapshotsDTO nodeStatusSnapshot : nodeStatusSnapshots) {
-            for (final StatusSnapshotDTO snapshotDto : nodeStatusSnapshot.getStatusSnapshots()) {
-                final StatusSnapshot snapshot = createSnapshot(snapshotDto, metricDescriptors);
-                final Date normalizedDate = normalizeStatusSnapshotDate(snapshot.getTimestamp(), componentStatusSnapshotMillis);
-
-                Map<String, StatusSnapshot> nodeToSnapshotMap = dateToNodeSnapshots.get(normalizedDate);
-                if (nodeToSnapshotMap == null) {
-                    nodeToSnapshotMap = new HashMap<>();
-                    dateToNodeSnapshots.put(normalizedDate, nodeToSnapshotMap);
-                }
-                nodeToSnapshotMap.put(nodeStatusSnapshot.getNodeId(), snapshot);
-            }
-        }
-
-        // aggregate the snapshots by (normalized) timestamp
-        final Map<Date, List<StatusSnapshot>> snapshotsToAggregate = new TreeMap<>();
-        for (final Map.Entry<Date, Map<String, StatusSnapshot>> entry : dateToNodeSnapshots.entrySet()) {
-            final Date normalizedDate = entry.getKey();
-            final Map<String, StatusSnapshot> nodeToSnapshot = entry.getValue();
-            final List<StatusSnapshot> snapshotsForTimestamp = new ArrayList<>(nodeToSnapshot.values());
-            snapshotsToAggregate.put(normalizedDate, snapshotsForTimestamp);
-        }
-
-        final List<StatusSnapshotDTO> aggregatedSnapshots = aggregate(snapshotsToAggregate);
-        return aggregatedSnapshots;
     }
